refactor(map): split region map render parts and add nesting lint (Refs #1746)
- Replace monolithic render mixin with four part files (orchestrator, core,
  political, markers) plus warp glow geometry helper and tests.
- Flatten warp-zone painting via _paintWarpZoneEdge and _warpTileCoordsForZones;
  document defense-in-depth warp-cell guard.
- Reduce nesting in _getDominantAdjacentLandBase with a tally helper.
- Add repo.control_flow_nesting_depth (guard-aware AST) and shrink-only
  allowlist for legacy depth>=4 symbols; document in SPEC/program.
- Extend repo-lint manifest and ct_repo_lint_lib in-process wiring.

Made-with: Cursor
