--- app/lib/features/game/flame/region_map_component_render_mixin.dart
@@ -1,1633 +1,0 @@
-part of 'region_map_component.dart';
-
-extension _CtRegionMapRenderExtension on CtRegionMapComponent {
-  TileVisibility _visibilityForTerrain(CellViewData cell) {
-    if (visibilityMode != CtMapVisibilityMode.playerConstrained) {
-      return cell.visibility;
-    }
-    if (_isUnderFleetRevealHalo(cell.x, cell.y)) {
-      return TileVisibility.visible;
-    }
-    return cell.visibility;
-  }
-
-  bool _isUnderFleetRevealHalo(int x, int y) {
-    for (final m in region.fleetTileMarkers) {
-      if (!m.applyFleetRevealHalo) {
-        continue;
-      }
-      if (math.max((x - m.x).abs(), (y - m.y).abs()) <= 2) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  void _renderRegionMap(Canvas canvas) {
-    _paintTiles(canvas);
-    if (showProvinceOwnershipTint) {
-      _paintGreatPowerLandOwnershipTint(canvas);
-    }
-    _paintOverlay(canvas);
-    if (showProvinceOverlay) {
-      _paintProvinceBorders(canvas);
-    }
-    if (_hoveredProvinceId != null) {
-      _paintHoveredProvinceGlow(canvas);
-    }
-    if (showPoliticalOverlay && showProvinceOverlay) {
-      _paintFactionBorders(canvas);
-    }
-    if (showProvinceNamesLayer) {
-      _paintProvinceNames(canvas);
-    }
-    _paintCapitals(canvas);
-    _paintTowns(canvas);
-    _paintWarpZones(canvas);
-    _paintCivilianTileMarkers(canvas);
-    _paintFleetTileMarkers(canvas);
-    if (_hoveredTileX != null && _hoveredTileY != null) {
-      _paintSelector(canvas);
-    }
-    if (selectedTileKey != null) {
-      _paintSelectedTile(canvas);
-    }
-    if (secondaryHighlightTileKey != null) {
-      _paintSecondaryHighlightTile(canvas);
-    }
-    if (validTileKeys != null && validTileKeys!.isNotEmpty) {
-      _paintValidTilesGlow(canvas);
-    }
-  }
-
-  void _paintCivilianTileMarkers(Canvas canvas) {
-    if (region.civilianTileMarkers.isEmpty) return;
-    for (final marker in region.civilianTileMarkers) {
-      if (marker.x < 0 ||
-          marker.x >= region.width ||
-          marker.y < 0 ||
-          marker.y >= region.height) {
-        continue;
-      }
-      final cell = region.cellAt(marker.x, marker.y);
-      if (regionMapSkipPointMarkerOnCell(
-        playerConstrainedVisibility:
-            visibilityMode == CtMapVisibilityMode.playerConstrained,
-        cellVisibility: _visibilityForTerrain(cell),
-      )) {
-        continue;
-      }
-      final left = marker.x * cellSize;
-      final top = marker.y * cellSize;
-      final selected = selectedCivilianTileKey == marker.tileKey;
-      final blinkAlpha = selected
-          ? _kHoveredProvinceGlowOpacityMid +
-                _kHoveredProvinceGlowOpacityAmplitude *
-                    math.sin(
-                      _hoverAnimationT * _kHoveredProvinceGlowAngularFrequency,
-                    )
-          : 1.0;
-
-      final icon = civilianIconCache.getIcon(
-        unitType: marker.representativeUnitType,
-        grayscale: marker.representativeIsAssigned,
-      );
-      if (icon != null) {
-        final srcRect = Rect.fromLTWH(
-          0,
-          0,
-          CivilianIconCache.iconSize,
-          CivilianIconCache.iconSize,
-        );
-        final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-        final paint = Paint();
-        final alpha = blinkAlpha.clamp(0.35, 1.0);
-        if (marker.representativeIsAssigned) {
-          paint.colorFilter = const ColorFilter.matrix(<double>[
-            0.2126,
-            0.7152,
-            0.0722,
-            0,
-            0,
-            0.2126,
-            0.7152,
-            0.0722,
-            0,
-            0,
-            0.2126,
-            0.7152,
-            0.0722,
-            0,
-            0,
-            0,
-            0,
-            0,
-            1,
-            0,
-          ]);
-        }
-        paint.color = Colors.white.withValues(alpha: alpha);
-        canvas.drawImageRect(icon, srcRect, dstRect, paint);
-      } else {
-        final fallback = Paint()
-          ..style = PaintingStyle.fill
-          ..color = const Color(
-            0xFF495057,
-          ).withValues(alpha: blinkAlpha.clamp(0.35, 1.0));
-        canvas.drawRect(Rect.fromLTWH(left, top, cellSize, cellSize), fallback);
-      }
-
-      if (marker.stackCount <= 1) continue;
-      final badgeRadius = math.max(7.0, cellSize * 0.2);
-      final badgeCx = left + cellSize - badgeRadius;
-      final badgeCy = top + cellSize - badgeRadius;
-      final badgeFill = Paint()
-        ..style = PaintingStyle.fill
-        ..color = Colors.black.withValues(alpha: blinkAlpha.clamp(0.35, 1.0));
-      canvas.drawCircle(Offset(badgeCx, badgeCy), badgeRadius, badgeFill);
-      final badgeText = TextPainter(
-        text: TextSpan(
-          text: '${marker.stackCount}',
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: math.max(8.0, cellSize * 0.24),
-            fontWeight: FontWeight.w700,
-          ),
-        ),
-        textDirection: TextDirection.ltr,
-      )..layout();
-      badgeText.paint(
-        canvas,
-        Offset(badgeCx - badgeText.width / 2, badgeCy - badgeText.height / 2),
-      );
-    }
-  }
-
-  void _paintFleetTileMarkers(Canvas canvas) {
-    if (region.fleetTileMarkers.isEmpty) {
-      return;
-    }
-    for (final marker in region.fleetTileMarkers) {
-      if (marker.x < 0 ||
-          marker.x >= region.width ||
-          marker.y < 0 ||
-          marker.y >= region.height) {
-        continue;
-      }
-      final left = marker.x * cellSize;
-      final top = marker.y * cellSize;
-      final icon = fleetIconCache.getIcon();
-      if (icon != null) {
-        final srcRect = Rect.fromLTWH(
-          0,
-          0,
-          FleetIconCache.iconSize,
-          FleetIconCache.iconSize,
-        );
-        final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-        final paint = Paint();
-        if (marker.renderGrayscale) {
-          paint.colorFilter = const ColorFilter.matrix(<double>[
-            0.2126,
-            0.7152,
-            0.0722,
-            0,
-            0,
-            0.2126,
-            0.7152,
-            0.0722,
-            0,
-            0,
-            0.2126,
-            0.7152,
-            0.0722,
-            0,
-            0,
-            0,
-            0,
-            0,
-            1,
-            0,
-          ]);
-        }
-        canvas.drawImageRect(icon, srcRect, dstRect, paint);
-      } else {
-        final fallback = Paint()
-          ..style = PaintingStyle.fill
-          ..color = const Color(0xFF1c3d5a);
-        canvas.drawRect(Rect.fromLTWH(left, top, cellSize, cellSize), fallback);
-      }
-
-      if (marker.stackCount <= 1) {
-        continue;
-      }
-      final badgeRadius = math.max(7.0, cellSize * 0.2);
-      final badgeCx = left + cellSize - badgeRadius;
-      final badgeCy = top + cellSize - badgeRadius;
-      final badgeFill = Paint()
-        ..style = PaintingStyle.fill
-        ..color = Colors.black;
-      canvas.drawCircle(Offset(badgeCx, badgeCy), badgeRadius, badgeFill);
-      final badgeText = TextPainter(
-        text: TextSpan(
-          text: '${marker.stackCount}',
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: math.max(8.0, cellSize * 0.24),
-            fontWeight: FontWeight.w700,
-          ),
-        ),
-        textDirection: TextDirection.ltr,
-      )..layout();
-      badgeText.paint(
-        canvas,
-        Offset(badgeCx - badgeText.width / 2, badgeCy - badgeText.height / 2),
-      );
-    }
-  }
-
-  void _paintValidTilesGlow(Canvas canvas) {
-    final keys = validTileKeys!;
-    final t = _hoverAnimationT;
-    final opacity =
-        _kValidWorkTargetGlowOpacityBase +
-        _kValidWorkTargetGlowOpacityAmplitude *
-            (_kSinNormalizedMid +
-                _kSinNormalizedMid *
-                    math.sin(t * _kValidWorkTargetGlowTimeScale));
-    final paint = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kMapValidTileTargetStrokeWidth
-      ..color = _kValidWorkTargetStrokeYellow.withValues(alpha: opacity);
-    for (var y = 0; y < region.height; y++) {
-      for (var x = 0; x < region.width; x++) {
-        final cell = region.cellAt(x, y);
-        final tileKey = '${region.regionId}|${cell.regionCellId}|$x|$y';
-        if (!keys.contains(tileKey)) continue;
-        final left = x * cellSize;
-        final top = y * cellSize;
-        canvas.drawRect(Rect.fromLTWH(left, top, cellSize, cellSize), paint);
-      }
-    }
-  }
-
-  void _paintSelectedTile(Canvas canvas) {
-    _paintTileOutlineRing(
-      canvas,
-      tileKey: selectedTileKey!,
-      color: _kMapSelectedHighlightOrange,
-      strokeWidth: kMapSelectedTileStrokeWidth,
-    );
-  }
-
-  void _paintSecondaryHighlightTile(Canvas canvas) {
-    _paintTileOutlineRing(
-      canvas,
-      tileKey: secondaryHighlightTileKey!,
-      color: _kMapSecondarySelectionCyan,
-      strokeWidth: kMapSecondaryHighlightStrokeWidth,
-    );
-  }
-
-  void _paintTileOutlineRing(
-    Canvas canvas, {
-    required String tileKey,
-    required Color color,
-    required double strokeWidth,
-  }) {
-    final parts = tileKey.split('|');
-    if (parts.length < 4) return;
-    if (parts[0] != region.regionId) return;
-    final x = int.tryParse(parts[2]);
-    final y = int.tryParse(parts[3]);
-    if (x == null || y == null) return;
-    if (x < 0 || x >= region.width || y < 0 || y >= region.height) return;
-    final left = x * cellSize;
-    final top = y * cellSize;
-    final rect = Rect.fromLTWH(left, top, cellSize, cellSize);
-    final paint = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = strokeWidth
-      ..color = color;
-    canvas.drawRect(rect, paint);
-  }
-
-  void _ensureProvinceLabelCache() {
-    if (identical(_provinceLabelsRegionRef, region) &&
-        _provinceLabelsCellSize == cellSize &&
-        _provinceLabelsVisibilityMode == visibilityMode &&
-        _provinceLabelsCached != null) {
-      return;
-    }
-    _provinceLabelsRegionRef = region;
-    _provinceLabelsCellSize = cellSize;
-    _provinceLabelsVisibilityMode = visibilityMode;
-    _provinceLabelsCached = _computeProvinceLabels();
-  }
-
-  Color _provinceNamePlateColor({
-    required String prefixedProvinceId,
-    required List<CellViewData> qualifyingLandCells,
-  }) {
-    final rgb = resolveProvinceLabelPlateTintRgb(
-      prefixedProvinceId: prefixedProvinceId,
-      qualifyingLandCells: qualifyingLandCells,
-      region: region,
-      honorUnrevealedTiles:
-          visibilityMode == CtMapVisibilityMode.playerConstrained,
-    );
-    if (rgb == null) {
-      return _provinceLabelPlateColor;
-    }
-    return Color.fromRGBO(rgb.$1, rgb.$2, rgb.$3, kProvinceLabelPlateTintAlpha);
-  }
-
-  List<
-    ({
-      double cx,
-      double cy,
-      String text,
-      String provinceId,
-      Color plateColor,
-      bool isCapital,
-    })
-  >
-  _computeProvinceLabels() {
-    final byLocalId = <String, List<CellViewData>>{};
-    for (final cell in region.cells) {
-      if (cell.isSea) continue;
-      if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-          _visibilityForTerrain(cell) == TileVisibility.unrevealed) {
-        continue;
-      }
-      byLocalId.putIfAbsent(cell.regionCellId, () => []).add(cell);
-    }
-    final out =
-        <
-          ({
-            double cx,
-            double cy,
-            String text,
-            String provinceId,
-            Color plateColor,
-            bool isCapital,
-          })
-        >[];
-    final capitalProvinceIds = <String>{};
-    for (final cap in region.capitalMarkers) {
-      if (cap.x < 0 ||
-          cap.x >= region.width ||
-          cap.y < 0 ||
-          cap.y >= region.height) {
-        continue;
-      }
-      final capCell = region.cellAt(cap.x, cap.y);
-      if (capCell.isSea) {
-        continue;
-      }
-      capitalProvinceIds.add('${region.regionId}|${capCell.regionCellId}');
-    }
-    for (final e in byLocalId.entries) {
-      final cells = e.value;
-      if (cells.isEmpty) continue;
-      var sx = 0.0;
-      var sy = 0.0;
-      for (final c in cells) {
-        sx += (c.x + 0.5) * cellSize;
-        sy += (c.y + 0.5) * cellSize;
-      }
-      final n = cells.length;
-      final cx = sx / n;
-      final cy = sy / n;
-      String? name;
-      for (final c in cells) {
-        final dn = c.provinceDisplayName;
-        if (dn != null && dn.isNotEmpty) {
-          name = dn;
-          break;
-        }
-      }
-      final text = name ?? e.key;
-      final prefixedId = '${region.regionId}|${e.key}';
-      out.add((
-        cx: cx,
-        cy: cy,
-        text: text,
-        provinceId: prefixedId,
-        plateColor: _provinceNamePlateColor(
-          prefixedProvinceId: prefixedId,
-          qualifyingLandCells: cells,
-        ),
-        isCapital: capitalProvinceIds.contains(prefixedId),
-      ));
-    }
-    return out;
-  }
-
-  void _paintProvinceNames(Canvas canvas) {
-    _ensureProvinceLabelCache();
-    final labels = _provinceLabelsCached;
-    if (labels == null || labels.isEmpty) {
-      return;
-    }
-
-    final invZ = 1.0 / cameraZoom.clamp(0.25, 4.0);
-    const textStyle = TextStyle(
-      color: Colors.white,
-      fontSize: _provinceLabelFontSizePx,
-      fontWeight: FontWeight.w600,
-      shadows: <Shadow>[
-        Shadow(
-          blurRadius: 2,
-          color: _kProvinceLabelShadowColor,
-          offset: Offset(0.5, 0.5),
-        ),
-      ],
-    );
-    for (final item in labels) {
-      final platePaint = Paint()..color = item.plateColor;
-      final presence = region.provinceUnitPresenceByProvinceId[item.provinceId];
-      final hasCapitalIcon = item.isCapital;
-      const capitalIconIds = <String>[_provinceLabelCapitalIconId];
-      final presenceIconIds = resolveProvinceLabelPresenceIconIds(presence);
-      final presenceIconCount = presenceIconIds.length;
-      final presenceIconsWidth = presenceIconCount > 0
-          ? (presenceIconCount * _provinceLabelIconRenderedPx) +
-                ((presenceIconCount - 1) * _provinceLabelIconGapPx)
-          : 0.0;
-      final capitalInlineWidth = hasCapitalIcon
-          ? _provinceLabelIconRenderedPx + _provinceLabelTextIconGapPx
-          : 0.0;
-      final shouldEllipsize = shouldEllipsizeProvinceLabelText(
-        isCapital: item.isCapital,
-      );
-      final tp =
-          TextPainter(
-            text: TextSpan(text: item.text, style: textStyle),
-            textDirection: TextDirection.ltr,
-            maxLines: shouldEllipsize ? 3 : null,
-            ellipsis: shouldEllipsize ? '…' : null,
-          )..layout(
-            maxWidth: shouldEllipsize
-                ? _provinceLabelMaxWidthPx
-                : double.infinity,
-          );
-
-      final tw = tp.width;
-      final th = tp.height;
-      final singleLinePresenceWidth = presenceIconCount > 0
-          ? _provinceLabelTextIconGapPx + presenceIconsWidth
-          : 0.0;
-      final singleLineContentWidth =
-          capitalInlineWidth + tw + singleLinePresenceWidth;
-      final wrapIconsToSecondLine = shouldWrapProvinceLabelPresenceIcons(
-        textWidthPx: tw,
-        iconCount: presenceIconCount,
-      );
-      final lineOneWidth = capitalInlineWidth + tw;
-      final lineOneHeight = hasCapitalIcon
-          ? math.max(th, _provinceLabelIconRenderedPx)
-          : th;
-      final contentWidth = wrapIconsToSecondLine
-          ? math.max(lineOneWidth, presenceIconsWidth)
-          : singleLineContentWidth;
-      final contentHeight = wrapIconsToSecondLine
-          ? lineOneHeight +
-                _provinceLabelTextIconGapPx +
-                _provinceLabelIconRenderedPx
-          : math.max(
-              lineOneHeight,
-              presenceIconCount > 0 ? _provinceLabelIconRenderedPx : 0,
-            );
-      const pad = _provinceLabelPlatePaddingPx;
-      final bw = contentWidth + pad * 2;
-      final bh = contentHeight + pad * 2;
-
-      canvas.save();
-      canvas.translate(item.cx, item.cy);
-      canvas.scale(invZ);
-      final rect = RRect.fromRectAndRadius(
-        Rect.fromCenter(center: Offset.zero, width: bw, height: bh),
-        const Radius.circular(4),
-      );
-      canvas.drawRRect(rect, platePaint);
-
-      if (wrapIconsToSecondLine) {
-        final lineOneLeft = -lineOneWidth / 2;
-        final lineOneTop = -contentHeight / 2;
-        if (hasCapitalIcon) {
-          _paintProvinceLabelIconsRow(
-            canvas: canvas,
-            iconIds: capitalIconIds,
-            left: lineOneLeft,
-            top:
-                lineOneTop + (lineOneHeight - _provinceLabelIconRenderedPx) / 2,
-            rowWidth: _provinceLabelIconRenderedPx,
-          );
-        }
-        tp.paint(
-          canvas,
-          Offset(
-            lineOneLeft + capitalInlineWidth,
-            lineOneTop + (lineOneHeight - th) / 2,
-          ),
-        );
-        final iconTop =
-            lineOneTop + lineOneHeight + _provinceLabelTextIconGapPx;
-        _paintProvinceLabelIconsRow(
-          canvas: canvas,
-          iconIds: presenceIconIds,
-          top: iconTop,
-          rowWidth: presenceIconsWidth,
-        );
-      } else {
-        final rowLeft = -singleLineContentWidth / 2;
-        final rowTop = -contentHeight / 2;
-        if (hasCapitalIcon) {
-          _paintProvinceLabelIconsRow(
-            canvas: canvas,
-            iconIds: capitalIconIds,
-            left: rowLeft,
-            top: rowTop + (contentHeight - _provinceLabelIconRenderedPx) / 2,
-            rowWidth: _provinceLabelIconRenderedPx,
-          );
-        }
-        final textLeft = rowLeft + capitalInlineWidth;
-        final textTop = rowTop + (contentHeight - th) / 2;
-        tp.paint(canvas, Offset(textLeft, textTop));
-        if (presenceIconCount > 0) {
-          final iconLeft = textLeft + tw + _provinceLabelTextIconGapPx;
-          _paintProvinceLabelIconsRow(
-            canvas: canvas,
-            iconIds: presenceIconIds,
-            left: iconLeft,
-            top: rowTop + (contentHeight - _provinceLabelIconRenderedPx) / 2,
-            rowWidth: presenceIconsWidth,
-          );
-        }
-      }
-      canvas.restore();
-    }
-  }
-
-  void _paintProvinceLabelIconsRow({
-    required Canvas canvas,
-    required List<String> iconIds,
-    required double top,
-    required double rowWidth,
-    double? left,
-  }) {
-    if (iconIds.isEmpty) return;
-    final srcRect = Rect.fromLTWH(
-      0,
-      0,
-      ProvinceLabelIconCache.iconSize,
-      ProvinceLabelIconCache.iconSize,
-    );
-    var x = left ?? (-rowWidth / 2);
-    for (final iconId in iconIds) {
-      final icon = provinceLabelIconCache.getIcon(iconId);
-      if (icon == null) {
-        x += _provinceLabelIconRenderedPx + _provinceLabelIconGapPx;
-        continue;
-      }
-      final dstRect = Rect.fromLTWH(
-        x,
-        top,
-        _provinceLabelIconRenderedPx,
-        _provinceLabelIconRenderedPx,
-      );
-      canvas.drawImageRect(icon, srcRect, dstRect, Paint());
-      x += _provinceLabelIconRenderedPx + _provinceLabelIconGapPx;
-    }
-  }
-
-  void _paintTiles(Canvas canvas) {
-    if (!terrainTilesetCache.isLoaded) {
-      return;
-    }
-    _paintTilesWithTilesets(canvas);
-  }
-
-  void _paintTilesWithTilesets(Canvas canvas) {
-    for (final cell in region.cells) {
-      if (cell.isSea) {
-        _paintSeaCell(canvas, cell);
-      }
-    }
-
-    for (final cell in region.cells) {
-      if (!cell.isSea) {
-        _paintLandBaseCell(canvas, cell);
-      }
-    }
-
-    for (final cell in region.cells) {
-      if (!cell.isSea &&
-          cell.terrainType != null &&
-          _isFeatureTerrain(cell.terrainType!)) {
-        _paintFeatureCell(canvas, cell);
-      }
-    }
-  }
-
-  void _paintSeaCell(Canvas canvas, CellViewData cell) {
-    final left = cell.x * cellSize;
-    final top = cell.y * cellSize;
-
-    if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-        _visibilityForTerrain(cell) == TileVisibility.unrevealed) {
-      final paint = Paint()..color = Colors.black;
-      canvas.drawRect(Rect.fromLTWH(left, top, cellSize, cellSize), paint);
-      return;
-    }
-
-    final landCorner = _getCoastlineCornerValues(cell.x, cell.y);
-    final dominantLandType = _getDominantAdjacentLandBase(
-      cell.x,
-      cell.y,
-      _getCellAt,
-    );
-    final tileset = dominantLandType == TerrainType.desert
-        ? terrainTilesetCache.getSeaDesertTileset()
-        : terrainTilesetCache.getSeaPlainsTileset();
-
-    if (tileset == null) {
-      throw StateError(
-        'Sea tileset is null for dominantLandType=$dominantLandType - '
-        'terrain tileset failed to load',
-      );
-    }
-
-    if (landCorner.same) {
-      final seaBaseTileId = tileset.lowerBaseTileId;
-      final tile = seaBaseTileId != null
-          ? tileset.findTileById(seaBaseTileId)
-          : tileset.findTile(nw: false, ne: false, sw: false, se: false);
-      if (tile != null) {
-        final srcRect = tile.boundingBox;
-        final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-        canvas.drawImageRect(tileset.image, srcRect, dstRect, Paint());
-        if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-            _visibilityForTerrain(cell) == TileVisibility.fogged) {
-          canvas.drawRect(
-            dstRect,
-            Paint()..color = Color.fromRGBO(0, 0, 0, _fogOverlayOpacity),
-          );
-        }
-      }
-      return;
-    }
-
-    _drawTile(
-      canvas,
-      tileset,
-      left,
-      top,
-      nw: landCorner.nw,
-      ne: landCorner.ne,
-      sw: landCorner.sw,
-      se: landCorner.se,
-      cell: cell,
-    );
-  }
-
-  void _paintLandBaseCell(Canvas canvas, CellViewData cell) {
-    final left = cell.x * cellSize;
-    final top = cell.y * cellSize;
-
-    if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-        _visibilityForTerrain(cell) == TileVisibility.unrevealed) {
-      final paint = Paint()..color = Colors.black;
-      canvas.drawRect(Rect.fromLTWH(left, top, cellSize, cellSize), paint);
-      return;
-    }
-
-    final terrain = cell.terrainType;
-    if (terrain == null) {
-      throw StateError('Cell has no terrain type: $cell');
-    }
-    _paintLandBaseTile(canvas, cell);
-  }
-
-  /// Wang interior upper-base tile for land: `sea_plains` for plains (and
-  /// feature terrains on plains) or `sea_desert` for desert. Used for L1 cells
-  /// and for the opaque grass under transparent `tile_plains_*` overlays.
-  void _drawLandInteriorUpperBaseForTerrain(
-    Canvas canvas, {
-    required TerrainType landTerrain,
-    required Rect dstRect,
-    required Paint paint,
-  }) {
-    final interiorTileset = landTerrain == TerrainType.desert
-        ? terrainTilesetCache.getSeaDesertTileset()
-        : terrainTilesetCache.getSeaPlainsTileset();
-    if (interiorTileset == null) {
-      throw StateError(
-        'Interior tileset is null for terrain=$landTerrain - '
-        'terrain tileset failed to load',
-      );
-    }
-    final tile = interiorTileset.upperBaseTileId != null
-        ? interiorTileset.findTileById(interiorTileset.upperBaseTileId!)
-        : null;
-    if (tile == null) {
-      throw StateError(
-        'Base tile not found for terrain=$landTerrain - '
-        'upperBaseTileId=${interiorTileset.upperBaseTileId}',
-      );
-    }
-    canvas.drawImageRect(
-      interiorTileset.image,
-      tile.boundingBox,
-      dstRect,
-      paint,
-    );
-  }
-
-  Paint _landBaseImagePaint({
-    required TerrainType terrain,
-    required TileVisibility tileVisibility,
-  }) {
-    final paint = Paint();
-    if (shouldApplyFogToLandBase(
-      visibilityMode: visibilityMode,
-      tileVisibility: tileVisibility,
-      terrain: terrain,
-    )) {
-      paint.colorFilter = ColorFilter.mode(
-        Color.fromRGBO(0, 0, 0, _fogOverlayOpacity),
-        BlendMode.darken,
-      );
-    }
-    return paint;
-  }
-
-  void _paintLandBaseTile(Canvas canvas, CellViewData cell) {
-    final left = cell.x * cellSize;
-    final top = cell.y * cellSize;
-    final terrainNullable = cell.terrainType;
-    if (terrainNullable == null) {
-      throw StateError('Cell has no terrain type: $cell');
-    }
-    final terrain = terrainNullable;
-
-    final isPlains =
-        terrain == TerrainType.plains || _isFeatureTerrain(terrain);
-    final isDesert = terrain == TerrainType.desert;
-
-    final nearDesertCorner = _getCornerValues(
-      cell.x,
-      cell.y,
-      (c) => !c.isSea && c.terrainType == TerrainType.desert,
-    );
-    final nearPlainsCorner = _getCornerValues(
-      cell.x,
-      cell.y,
-      (c) =>
-          !c.isSea &&
-          (c.terrainType == TerrainType.plains ||
-              (c.terrainType != null && _isFeatureTerrain(c.terrainType!))),
-    );
-
-    if (isPlains && !nearDesertCorner.same && nearDesertCorner.value) {
-      final tileset = terrainTilesetCache.getPlainsDesertTileset();
-      if (tileset == null) {
-        throw StateError(
-          'Plains desert tileset is null - terrain tileset failed to load',
-        );
-      }
-      _drawTile(
-        canvas,
-        tileset,
-        left,
-        top,
-        nw: nearDesertCorner.nw,
-        ne: nearDesertCorner.ne,
-        sw: nearDesertCorner.sw,
-        se: nearDesertCorner.se,
-        cell: cell,
-      );
-      return;
-    }
-
-    if (isDesert && !nearPlainsCorner.same && nearPlainsCorner.value) {
-      final tileset = terrainTilesetCache.getPlainsDesertTileset();
-      if (tileset == null) {
-        throw StateError(
-          'Plains desert tileset is null - terrain tileset failed to load',
-        );
-      }
-      _drawTile(
-        canvas,
-        tileset,
-        left,
-        top,
-        nw: !nearPlainsCorner.nw,
-        ne: !nearPlainsCorner.ne,
-        sw: !nearPlainsCorner.sw,
-        se: !nearPlainsCorner.se,
-        cell: cell,
-      );
-      return;
-    }
-
-    if (terrain == TerrainType.plains) {
-      final plainsVariantKey = landInteriorPlainsVariantTileKey(cell);
-      if (plainsVariantKey != null) {
-        final standaloneTile = terrainTilesetCache.getStandaloneTileByKey(
-          plainsVariantKey,
-        );
-        if (standaloneTile == null) {
-          throw StateError(
-            'Missing required plains terrain variant tile: $plainsVariantKey',
-          );
-        }
-        final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-        _drawLandInteriorUpperBaseForTerrain(
-          canvas,
-          landTerrain: TerrainType.plains,
-          dstRect: dstRect,
-          paint: Paint(),
-        );
-        final overlayPaint = Paint();
-        if (shouldApplyFogToInteriorPlainsVariantOverlay(
-          visibilityMode: visibilityMode,
-          tileVisibility: _visibilityForTerrain(cell),
-        )) {
-          overlayPaint.colorFilter = ColorFilter.mode(
-            Color.fromRGBO(0, 0, 0, _fogOverlayOpacity),
-            BlendMode.darken,
-          );
-        }
-        final srcRect = Rect.fromLTWH(
-          0,
-          0,
-          standaloneTile.image.width.toDouble(),
-          standaloneTile.image.height.toDouble(),
-        );
-        canvas.drawImageRect(
-          standaloneTile.image,
-          srcRect,
-          dstRect,
-          overlayPaint,
-        );
-        return;
-      }
-    }
-
-    final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-    final paint = _landBaseImagePaint(
-      terrain: terrain,
-      tileVisibility: _visibilityForTerrain(cell),
-    );
-    _drawLandInteriorUpperBaseForTerrain(
-      canvas,
-      landTerrain: terrain,
-      dstRect: dstRect,
-      paint: paint,
-    );
-  }
-
-  void _paintFeatureCell(Canvas canvas, CellViewData cell) {
-    final left = cell.x * cellSize;
-    final top = cell.y * cellSize;
-
-    if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-        _visibilityForTerrain(cell) == TileVisibility.unrevealed) {
-      return;
-    }
-
-    final terrain = cell.terrainType;
-    if (terrain == null || !_isFeatureTerrain(terrain)) return;
-
-    final overlayTileKey = featureOverlayTileKey(
-      terrain: terrain,
-      resourceId: cell.resourceId,
-      improvementLevel: cell.improvementLevel,
-    );
-    final standaloneTile =
-        terrainTilesetCache.getStandaloneTileByKey(overlayTileKey) ??
-        terrainTilesetCache.getStandaloneTile(terrain);
-    if (standaloneTile == null) {
-      _log.w(
-        'Feature overlay tile missing for key=$overlayTileKey terrain=$terrain; '
-        'skipping feature overlay for cell (${cell.x}, ${cell.y})',
-      );
-      return;
-    }
-
-    final paint = Paint();
-    if (shouldApplyFogToFeatureOverlay(
-      visibilityMode: visibilityMode,
-      tileVisibility: _visibilityForTerrain(cell),
-      terrain: terrain,
-    )) {
-      paint.colorFilter = ColorFilter.mode(
-        Color.fromRGBO(0, 0, 0, _fogOverlayOpacity),
-        BlendMode.darken,
-      );
-    }
-
-    final srcRect = Rect.fromLTWH(
-      0,
-      0,
-      standaloneTile.image.width.toDouble(),
-      standaloneTile.image.height.toDouble(),
-    );
-    final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-    canvas.drawImageRect(standaloneTile.image, srcRect, dstRect, paint);
-  }
-
-  _CornerValues _getCornerValues(
-    int x,
-    int y,
-    bool Function(CellViewData) predicate,
-  ) {
-    final nwCell = _getCellAt(x - 1, y - 1);
-    final nCell = _getCellAt(x, y - 1);
-    final neCell = _getCellAt(x + 1, y - 1);
-    final wCell = _getCellAt(x - 1, y);
-    final cCell = _getCellAt(x, y);
-    final eCell = _getCellAt(x + 1, y);
-    final swCell = _getCellAt(x - 1, y + 1);
-    final sCell = _getCellAt(x, y + 1);
-    final seCell = _getCellAt(x + 1, y + 1);
-
-    bool test(CellViewData? c) => c != null && predicate(c);
-
-    final centerMatches = cCell != null && predicate(cCell);
-    final hasNW = centerMatches && (test(nwCell) || test(nCell) || test(wCell));
-    final hasNE = centerMatches && (test(neCell) || test(nCell) || test(eCell));
-    final hasSW = centerMatches && (test(swCell) || test(sCell) || test(wCell));
-    final hasSE = centerMatches && (test(seCell) || test(sCell) || test(eCell));
-
-    final allSame = (hasNW == hasNE && hasNE == hasSW && hasSW == hasSE);
-    final same = allSame && (!hasNW || centerMatches);
-    final value = hasNW;
-
-    return _CornerValues(
-      nw: hasNW,
-      ne: hasNE,
-      sw: hasSW,
-      se: hasSE,
-      same: same,
-      value: value,
-    );
-  }
-
-  _CornerValues _getCoastlineCornerValues(int x, int y) {
-    final nwCell = _getCellAt(x - 1, y - 1);
-    final nCell = _getCellAt(x, y - 1);
-    final neCell = _getCellAt(x + 1, y - 1);
-    final wCell = _getCellAt(x - 1, y);
-    final eCell = _getCellAt(x + 1, y);
-    final swCell = _getCellAt(x - 1, y + 1);
-    final sCell = _getCellAt(x, y + 1);
-    final seCell = _getCellAt(x + 1, y + 1);
-
-    bool isLand(CellViewData? c) => c != null && !c.isSea;
-
-    final hasNW = isLand(nwCell) || isLand(nCell) || isLand(wCell);
-    final hasNE = isLand(neCell) || isLand(nCell) || isLand(eCell);
-    final hasSW = isLand(swCell) || isLand(sCell) || isLand(wCell);
-    final hasSE = isLand(seCell) || isLand(sCell) || isLand(eCell);
-
-    final allSame = (hasNW == hasNE && hasNE == hasSW && hasSW == hasSE);
-    final same = allSame && !hasNW;
-
-    return _CornerValues(
-      nw: hasNW,
-      ne: hasNE,
-      sw: hasSW,
-      se: hasSE,
-      same: same,
-      value: hasNW,
-    );
-  }
-
-  bool _drawTile(
-    Canvas canvas,
-    WangTileset tileset,
-    double left,
-    double top, {
-    required bool nw,
-    required bool ne,
-    required bool sw,
-    required bool se,
-    required CellViewData cell,
-  }) {
-    final tile = tileset.findTile(nw: nw, ne: ne, sw: sw, se: se);
-    if (tile == null) {
-      _log.w(
-        'No tile found in ${tileset.name} for corners: NW=$nw, NE=$ne, SW=$sw, SE=$se',
-      );
-      return false;
-    }
-
-    final srcRect = tile.boundingBox;
-    final dstRect = Rect.fromLTWH(left, top, cellSize, cellSize);
-    canvas.drawImageRect(tileset.image, srcRect, dstRect, Paint());
-
-    if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-        _visibilityForTerrain(cell) == TileVisibility.fogged) {
-      canvas.drawRect(
-        dstRect,
-        Paint()..color = Color.fromRGBO(0, 0, 0, _fogOverlayOpacity),
-      );
-    }
-    return true;
-  }
-
-  CellViewData? _getCellAt(int x, int y) {
-    if (x < 0 || x >= region.width || y < 0 || y >= region.height) return null;
-    return region.cellAt(x, y);
-  }
-
-  String? _resourceIdForMapIcon(CellViewData cell) {
-    final raw = cell.resourceId;
-    if (raw == null) return null;
-    if (visibilityMode != CtMapVisibilityMode.playerConstrained) {
-      return raw;
-    }
-    final view = playerViewForResources;
-    if (view == null) {
-      throw StateError(
-        'CtRegionMapComponent: playerConstrained requires playerViewForResources',
-      );
-    }
-    final tileKey =
-        '${region.regionId}|${cell.regionCellId}|${cell.x}|${cell.y}';
-    return resourceIdVisibleInPlayerView(view, tileKey, raw);
-  }
-
-  Paint _resourceOverlayPaintForCell(CellViewData cell) {
-    final paint = Paint();
-    if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-        _visibilityForTerrain(cell) == TileVisibility.fogged) {
-      paint.colorFilter = ColorFilter.mode(
-        _kMapHoverSelectorIdle.withValues(
-          alpha: _kFoggedResourceIconModulateAlpha,
-        ),
-        BlendMode.modulate,
-      );
-    }
-    return paint;
-  }
-
-  void _paintOverlay(Canvas canvas) {
-    final showResources =
-        baseLayerDisplayMode != BaseLayerDisplayMode.terrainOnly;
-    final showExtractionDiscs = shouldShowExtractionUnitDiscs(
-      baseLayerDisplayMode: baseLayerDisplayMode,
-    );
-    final showImprovementLabels =
-        baseLayerDisplayMode ==
-            BaseLayerDisplayMode.terrainAndResourcesImprovementLabels ||
-        baseLayerDisplayMode ==
-            BaseLayerDisplayMode.terrainAndResourcesImprovementsRoads;
-    final showRoadLabels =
-        baseLayerDisplayMode ==
-        BaseLayerDisplayMode.terrainAndResourcesImprovementsRoads;
-
-    for (final cell in region.cells) {
-      if (cell.isSea) continue;
-      if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-          _visibilityForTerrain(cell) == TileVisibility.unrevealed) {
-        continue;
-      }
-
-      final resourceForIcon = _resourceIdForMapIcon(cell);
-      if (showResources && resourceForIcon != null) {
-        final icon = resourceIconCache.getIcon(resourceForIcon);
-        if (icon != null) {
-          final assetSize = ResourceIconCache.iconSize;
-          final displaySize = resourceIconDisplaySizePx(cellSize);
-          final tileLeft = cell.x * cellSize;
-          final tileTop = cell.y * cellSize;
-
-          final iconX = tileLeft;
-          final iconY = tileTop + cellSize - displaySize;
-
-          final dstRect = Rect.fromLTWH(iconX, iconY, displaySize, displaySize);
-          final srcRect = Rect.fromLTWH(0, 0, assetSize, assetSize);
-          final paint = _resourceOverlayPaintForCell(cell);
-          canvas.drawImageRect(icon, srcRect, dstRect, paint);
-          final extractionUnits = cell.resourceExtractionUnits ?? 0;
-          if (showExtractionDiscs && extractionUnits > 0) {
-            final discColor = discColorForResourceId(resourceForIcon);
-            final discPaint = _resourceOverlayPaintForCell(cell)
-              ..style = PaintingStyle.fill
-              ..color = discColor;
-            final centers = extractionDiscCentersForIconRect(
-              iconRect: dstRect,
-              units: extractionUnits,
-            );
-            for (final center in centers) {
-              canvas.drawCircle(center, _kExtractionDiscRadiusPx, discPaint);
-            }
-          }
-        }
-      }
-    }
-
-    if (showImprovementLabels) {
-      for (final cell in region.cells) {
-        if (cell.isSea) continue;
-        if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-            _visibilityForTerrain(cell) == TileVisibility.unrevealed) {
-          continue;
-        }
-        final imp = cell.improvementLevel ?? 0;
-        if (imp <= 0) continue;
-        _paintTileCornerLabel(canvas, cell, 'I$imp', alignEnd: false);
-      }
-    }
-
-    if (showRoadLabels) {
-      for (final cell in region.cells) {
-        if (cell.isSea) continue;
-        if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-            _visibilityForTerrain(cell) == TileVisibility.unrevealed) {
-          continue;
-        }
-        final road = cell.roadLevel ?? 0;
-        if (road <= 0) continue;
-        _paintTileCornerLabel(canvas, cell, 'R$road', alignEnd: true);
-      }
-    }
-  }
-
-  void _paintTileCornerLabel(
-    Canvas canvas,
-    CellViewData cell,
-    String text, {
-    required bool alignEnd,
-  }) {
-    final tileLeft = cell.x * cellSize;
-    final tileTop = cell.y * cellSize;
-    final pad = math.max(1.0, cellSize * 0.06);
-    final fontSize = math.max(8.0, cellSize * 0.25);
-    final textPainter = TextPainter(textDirection: TextDirection.ltr);
-    textPainter.text = TextSpan(
-      text: text,
-      style: TextStyle(
-        color: Colors.black,
-        fontSize: fontSize,
-        fontWeight: FontWeight.w600,
-      ),
-    );
-    textPainter.layout(maxWidth: cellSize - 2 * pad);
-    final y = tileTop + pad;
-    final x = alignEnd
-        ? tileLeft + cellSize - pad - textPainter.width
-        : tileLeft + pad;
-    textPainter.paint(canvas, Offset(x, y));
-  }
-
-  void _paintHoveredProvinceGlow(Canvas canvas) {
-    final t = _hoverAnimationT;
-    final opacity =
-        _kHoveredProvinceGlowOpacityMid +
-        _kHoveredProvinceGlowOpacityAmplitude *
-            math.sin(t * _kHoveredProvinceGlowAngularFrequency);
-    final coastInset = provinceOverlayLandSeaInsetPx(
-      cellSizePx: cellSize,
-      topologyStrokeWidth: kProvinceOverlayTopologyStrokeWidth,
-    );
-    final paint = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kProvinceOverlayHoverGlowStrokeWidth
-      ..color = _kMapHoverSelectorIdle.withValues(alpha: opacity);
-    final provinceId = _hoveredProvinceId!;
-    for (var y = 0; y < region.height; y++) {
-      for (var x = 0; x < region.width; x++) {
-        final cell = region.cellAt(x, y);
-        if (cell.regionCellId != provinceId) continue;
-        if (x + 1 < region.width) {
-          final right = region.cellAt(x + 1, y);
-          if (right.regionCellId != provinceId) {
-            if (regionMapDrawBoundaryBetweenAdjacentCells(
-              gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-              visibilityA: _visibilityForTerrain(cell),
-              visibilityB: _visibilityForTerrain(right),
-            )) {
-              final xEdge = verticalProvinceTopologyEdgeX(
-                left: cell,
-                right: right,
-                cellSizePx: cellSize,
-                leftTileX: x,
-                coastInsetPx: coastInset,
-              );
-              canvas.drawLine(
-                Offset(xEdge, y * cellSize),
-                Offset(xEdge, (y + 1) * cellSize),
-                paint,
-              );
-            }
-          }
-        }
-        if (y + 1 < region.height) {
-          final bottom = region.cellAt(x, y + 1);
-          if (bottom.regionCellId != provinceId) {
-            if (regionMapDrawBoundaryBetweenAdjacentCells(
-              gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-              visibilityA: _visibilityForTerrain(cell),
-              visibilityB: _visibilityForTerrain(bottom),
-            )) {
-              final yEdge = horizontalProvinceTopologyEdgeY(
-                top: cell,
-                bottom: bottom,
-                cellSizePx: cellSize,
-                topTileY: y,
-                coastInsetPx: coastInset,
-              );
-              canvas.drawLine(
-                Offset(x * cellSize, yEdge),
-                Offset((x + 1) * cellSize, yEdge),
-                paint,
-              );
-            }
-          }
-        }
-      }
-    }
-  }
-
-  void _paintSelector(Canvas canvas) {
-    final x = _hoveredTileX!;
-    final y = _hoveredTileY!;
-    final bounce =
-        _kHoverSelectorBounceBaseline +
-        _kHoverSelectorBounceAmplitude *
-            math.sin(_hoverAnimationT * _kHoveredProvinceGlowAngularFrequency);
-    final cx = x * cellSize + cellSize / 2;
-    final cy = y * cellSize + cellSize / 2;
-    final half = (cellSize / 2 - 2.0) * bounce;
-    final left = cx - half;
-    final top = cy - half;
-    final size = half * 2;
-    final rect = Rect.fromLTWH(left, top, size, size);
-    final color = (validTileKeys != null && validTileKeys!.isNotEmpty)
-        ? _kMapSelectedHighlightOrange
-        : _kMapHoverSelectorIdle;
-    final paint = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kMapHoverSelectorStrokeWidth
-      ..color = color;
-    canvas.drawRect(rect, paint);
-  }
-
-  void _paintGreatPowerLandOwnershipTint(Canvas canvas) {
-    paintGreatPowerOwnershipTintLayer(
-      canvas: canvas,
-      region: region,
-      cellSize: cellSize,
-      honorUnrevealedTiles:
-          visibilityMode == CtMapVisibilityMode.playerConstrained,
-    );
-  }
-
-  void _paintProvinceBorders(Canvas canvas) {
-    final coastInset = provinceOverlayLandSeaInsetPx(
-      cellSizePx: cellSize,
-      topologyStrokeWidth: kProvinceOverlayTopologyStrokeWidth,
-    );
-    final paint = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kProvinceOverlayTopologyStrokeWidth
-      ..color = _provinceBorderLandColor;
-    for (var y = 0; y < region.height; y++) {
-      for (var x = 0; x < region.width; x++) {
-        final cell = region.cellAt(x, y);
-        if (x + 1 < region.width) {
-          final right = region.cellAt(x + 1, y);
-          if (cell.regionCellId != right.regionCellId) {
-            if (regionMapDrawBoundaryBetweenAdjacentCells(
-              gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-              visibilityA: _visibilityForTerrain(cell),
-              visibilityB: _visibilityForTerrain(right),
-            )) {
-              paint.color = _provinceBorderColor(cell, right);
-              final xEdge = verticalProvinceTopologyEdgeX(
-                left: cell,
-                right: right,
-                cellSizePx: cellSize,
-                leftTileX: x,
-                coastInsetPx: coastInset,
-              );
-              canvas.drawLine(
-                Offset(xEdge, y * cellSize),
-                Offset(xEdge, (y + 1) * cellSize),
-                paint,
-              );
-            }
-          }
-        }
-        if (y + 1 < region.height) {
-          final bottom = region.cellAt(x, y + 1);
-          if (cell.regionCellId != bottom.regionCellId) {
-            if (regionMapDrawBoundaryBetweenAdjacentCells(
-              gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-              visibilityA: _visibilityForTerrain(cell),
-              visibilityB: _visibilityForTerrain(bottom),
-            )) {
-              paint.color = _provinceBorderColor(cell, bottom);
-              final yEdge = horizontalProvinceTopologyEdgeY(
-                top: cell,
-                bottom: bottom,
-                cellSizePx: cellSize,
-                topTileY: y,
-                coastInsetPx: coastInset,
-              );
-              canvas.drawLine(
-                Offset(x * cellSize, yEdge),
-                Offset((x + 1) * cellSize, yEdge),
-                paint,
-              );
-            }
-          }
-        }
-      }
-    }
-  }
-
-  Color _provinceBorderColor(CellViewData a, CellViewData b) {
-    final aIsSea = a.isSea;
-    final bIsSea = b.isSea;
-    if (aIsSea && bIsSea) return _provinceBorderSeaZoneColor;
-    if (!aIsSea && !bIsSea) return _provinceBorderLandColor;
-    return _provinceBorderSeaLandColor;
-  }
-
-  void _paintFactionBorders(Canvas canvas) {
-    final paint = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kProvinceOverlayPoliticalStrokeWidth
-      ..color = _kFactionPoliticalBorderColor;
-    for (var y = 0; y < region.height; y++) {
-      for (var x = 0; x < region.width; x++) {
-        final cell = region.cellAt(x, y);
-        if (cell.isSea) continue;
-        final owner = cell.ownerFactionId ?? '';
-        if (x + 1 < region.width) {
-          final right = region.cellAt(x + 1, y);
-          if (!right.isSea && (right.ownerFactionId ?? '') != owner) {
-            if (regionMapDrawBoundaryBetweenAdjacentCells(
-              gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-              visibilityA: _visibilityForTerrain(cell),
-              visibilityB: _visibilityForTerrain(right),
-            )) {
-              final xEdge = (x + 1) * cellSize;
-              canvas.drawLine(
-                Offset(xEdge, y * cellSize),
-                Offset(xEdge, (y + 1) * cellSize),
-                paint,
-              );
-            }
-          }
-        }
-        if (y + 1 < region.height) {
-          final bottom = region.cellAt(x, y + 1);
-          if (!bottom.isSea && (bottom.ownerFactionId ?? '') != owner) {
-            if (regionMapDrawBoundaryBetweenAdjacentCells(
-              gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-              visibilityA: _visibilityForTerrain(cell),
-              visibilityB: _visibilityForTerrain(bottom),
-            )) {
-              final yEdge = (y + 1) * cellSize;
-              canvas.drawLine(
-                Offset(x * cellSize, yEdge),
-                Offset((x + 1) * cellSize, yEdge),
-                paint,
-              );
-            }
-          }
-        }
-      }
-    }
-  }
-
-  void _paintCapitals(Canvas canvas) {
-    final fill = Paint()..style = PaintingStyle.fill;
-    final stroke = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kCapitalMarkerRingStrokeWidth
-      ..color = Colors.black;
-    for (final cap in region.capitalMarkers) {
-      final cell = region.cellAt(cap.x, cap.y);
-      if (regionMapSkipPointMarkerOnCell(
-        playerConstrainedVisibility:
-            visibilityMode == CtMapVisibilityMode.playerConstrained,
-        cellVisibility: _visibilityForTerrain(cell),
-      )) {
-        continue;
-      }
-      final cx = cap.x * cellSize + cellSize / 2;
-      final cy = cap.y * cellSize + cellSize / 2;
-      fill.color = _kMapSelectionGold;
-      canvas.drawCircle(Offset(cx, cy), 6, fill);
-      canvas.drawCircle(Offset(cx, cy), 6, stroke);
-    }
-  }
-
-  void _paintTowns(Canvas canvas) {
-    if (!townIconCache.isLoaded) return;
-
-    for (final town in region.townMarkers) {
-      final cell = region.cellAt(town.x, town.y);
-      if (regionMapSkipPointMarkerOnCell(
-        playerConstrainedVisibility:
-            visibilityMode == CtMapVisibilityMode.playerConstrained,
-        cellVisibility: _visibilityForTerrain(cell),
-      )) {
-        continue;
-      }
-      _paintTownIconAt(
-        canvas,
-        cell: cell,
-        cx: town.x,
-        cy: town.y,
-        icon: TownIconCache.townIconId,
-      );
-    }
-
-    for (final town in region.townMarkers) {
-      if (!town.isPort) continue;
-      final px = town.portIconX;
-      final py = town.portIconY;
-      if (px == null || py == null) continue;
-      final portCell = region.cellAt(px, py);
-      if (regionMapSkipPointMarkerOnCell(
-        playerConstrainedVisibility:
-            visibilityMode == CtMapVisibilityMode.playerConstrained,
-        cellVisibility: _visibilityForTerrain(portCell),
-      )) {
-        continue;
-      }
-      _paintTownIconAt(
-        canvas,
-        cell: portCell,
-        cx: px,
-        cy: py,
-        icon: TownIconCache.portIconId,
-      );
-    }
-  }
-
-  void _paintTownIconAt(
-    Canvas canvas, {
-    required CellViewData cell,
-    required int cx,
-    required int cy,
-    required String icon,
-  }) {
-    final uiImage = townIconCache.getIcon(icon);
-    if (uiImage == null) return;
-
-    final centerX = cx * cellSize + cellSize / 2;
-    final centerY = cy * cellSize + cellSize / 2;
-    final iconSize = icon == TownIconCache.portIconId
-        ? TownIconCache.portIconSize
-        : TownIconCache.townIconSize;
-    final halfIcon = iconSize / 2;
-
-    final srcRect = Rect.fromLTWH(
-      0,
-      0,
-      uiImage.width.toDouble(),
-      uiImage.height.toDouble(),
-    );
-    final dstRect = Rect.fromLTWH(
-      centerX - halfIcon,
-      centerY - halfIcon,
-      iconSize,
-      iconSize,
-    );
-
-    var paint = Paint();
-    if (visibilityMode == CtMapVisibilityMode.playerConstrained &&
-        _visibilityForTerrain(cell) == TileVisibility.fogged) {
-      paint.color = Color.fromRGBO(0, 0, 0, _fogOverlayOpacity);
-    }
-    canvas.drawImageRect(uiImage, srcRect, dstRect, paint);
-  }
-
-  void _paintWarpZones(Canvas canvas) {
-    final warpSeaZoneIds = region.warpMarkers.map((m) => m.seaZoneId).toSet();
-    if (warpSeaZoneIds.isEmpty) return;
-
-    final warpTiles = <(int x, int y)>{};
-    for (var y = 0; y < region.height; y++) {
-      for (var x = 0; x < region.width; x++) {
-        final cell = region.cellAt(x, y);
-        if (warpSeaZoneIds.contains(cell.regionCellId)) {
-          warpTiles.add((x, y));
-        }
-      }
-    }
-
-    final glowOuter = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kWarpZoneGlowOuterStrokeWidth
-      ..color = _kMapSelectionGold.withValues(alpha: _kWarpZoneOuterGlowAlpha);
-    final glowInner = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kWarpZoneGlowInnerStrokeWidth
-      ..color = _kWarpZoneInnerHighlight;
-
-    for (final (x, y) in warpTiles) {
-      final cell = region.cellAt(x, y);
-      if (!warpSeaZoneIds.contains(cell.regionCellId)) continue;
-
-      if (x + 1 < region.width) {
-        final right = region.cellAt(x + 1, y);
-        if (!warpSeaZoneIds.contains(right.regionCellId)) {
-          if (regionMapDrawBoundaryBetweenAdjacentCells(
-            gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-            visibilityA: _visibilityForTerrain(cell),
-            visibilityB: _visibilityForTerrain(right),
-          )) {
-            final xEdge = (x + 1) * cellSize;
-            canvas.drawLine(
-              Offset(xEdge, y * cellSize),
-              Offset(xEdge, (y + 1) * cellSize),
-              glowOuter,
-            );
-            canvas.drawLine(
-              Offset(xEdge, y * cellSize),
-              Offset(xEdge, (y + 1) * cellSize),
-              glowInner,
-            );
-          }
-        }
-      }
-      if (y + 1 < region.height) {
-        final bottom = region.cellAt(x, y + 1);
-        if (!warpSeaZoneIds.contains(bottom.regionCellId)) {
-          if (regionMapDrawBoundaryBetweenAdjacentCells(
-            gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-            visibilityA: _visibilityForTerrain(cell),
-            visibilityB: _visibilityForTerrain(bottom),
-          )) {
-            final yEdge = (y + 1) * cellSize;
-            canvas.drawLine(
-              Offset(x * cellSize, yEdge),
-              Offset((x + 1) * cellSize, yEdge),
-              glowOuter,
-            );
-            canvas.drawLine(
-              Offset(x * cellSize, yEdge),
-              Offset((x + 1) * cellSize, yEdge),
-              glowInner,
-            );
-          }
-        }
-      }
-      if (x > 0) {
-        final left = region.cellAt(x - 1, y);
-        if (!warpSeaZoneIds.contains(left.regionCellId)) {
-          if (regionMapDrawBoundaryBetweenAdjacentCells(
-            gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-            visibilityA: _visibilityForTerrain(cell),
-            visibilityB: _visibilityForTerrain(left),
-          )) {
-            final xEdge = x * cellSize;
-            canvas.drawLine(
-              Offset(xEdge, y * cellSize),
-              Offset(xEdge, (y + 1) * cellSize),
-              glowOuter,
-            );
-            canvas.drawLine(
-              Offset(xEdge, y * cellSize),
-              Offset(xEdge, (y + 1) * cellSize),
-              glowInner,
-            );
-          }
-        }
-      }
-      if (y > 0) {
-        final top = region.cellAt(x, y - 1);
-        if (!warpSeaZoneIds.contains(top.regionCellId)) {
-          if (regionMapDrawBoundaryBetweenAdjacentCells(
-            gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
-            visibilityA: _visibilityForTerrain(cell),
-            visibilityB: _visibilityForTerrain(top),
-          )) {
-            final yEdge = y * cellSize;
-            canvas.drawLine(
-              Offset(x * cellSize, yEdge),
-              Offset((x + 1) * cellSize, yEdge),
-              glowOuter,
-            );
-            canvas.drawLine(
-              Offset(x * cellSize, yEdge),
-              Offset((x + 1) * cellSize, yEdge),
-              glowInner,
-            );
-          }
-        }
-      }
-    }
-  }
-}
