+++ tool/ct_repo_lint_lib.dart
@@ -8,6 +8,7 @@ import 'check_app_hardcoded_ui_strings.dart';
 import 'check_asset_path_constants.dart';
 import 'check_canonical_province_tile_keys.dart';
 import 'check_civilian_unit_type_constants.dart';
+import 'check_control_flow_nesting_depth.dart';
 import 'check_custom_exceptions.dart';
 import 'check_disallowed_ast_patterns.dart';
 import 'check_repeated_magic_numbers.dart';
@@ -617,6 +618,8 @@ int? _tryRunDartRuleInProcess({
       return runCheckAssetPathConstants(repoRoot);
     case 'repo.disallowed_ast_patterns':
       return runCheckDisallowedAstPatterns(repoRoot);
+    case 'repo.control_flow_nesting_depth':
+      return runCheckControlFlowNestingDepth(repoRoot);
     case 'repo.repeated_magic_numbers':
       return runCheckRepeatedMagicNumbers(repoRoot);
     case 'repo.tech_id_constants':
